Migrate to Django-CMS

+++ src/open_inwoner/accounts/apps.py
@@ -1,12 +1,17 @@
 from io import StringIO
 
 from django.apps import AppConfig, apps
+from django.conf import settings
 from django.contrib.contenttypes.management import create_contenttypes
 from django.core.management import call_command
 from django.db.models.signals import post_migrate
 
 
 def update_admin_index(sender, **kwargs):
+    if "django_admin_index" not in settings.INSTALLED_APPS:
+        print("django_admin_index not installed, skipping update_admin_index()")
+        return
+
     from django_admin_index.models import AppGroup
 
     AppGroup.objects.all().delete()
@@ -18,7 +23,10 @@ def update_admin_index(sender, **kwargs):
         if app_config.name.startswith(project_name):
             create_contenttypes(app_config, verbosity=0)
 
-    call_command("loaddata", "django-admin-index", verbosity=0, stdout=StringIO())
+    try:
+        call_command("loaddata", "django-admin-index", verbosity=0, stdout=StringIO())
+    except:
+        print("Error: Unable to load django-admin-index fixture!")
 
 
 class AccountsConfig(AppConfig):
